requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
